Split Message types into separate classes
