+++ src/main/java/net/dv8tion/discord/bridge/endpoint/types/DiscordEndPoint.java
@@ -20,6 +20,8 @@ import net.dv8tion.discord.bridge.endpoint.EndPoint;
 import net.dv8tion.discord.bridge.endpoint.EndPointInfo;
 import net.dv8tion.discord.bridge.endpoint.EndPointMessage;
 import net.dv8tion.discord.bridge.endpoint.EndPointType;
+import net.dv8tion.discord.bridge.endpoint.messages.DiscordEndPointMessage;
+import net.dv8tion.discord.bridge.endpoint.messages.IrcEndPointMessage;
 import net.dv8tion.jda.entities.Guild;
 import net.dv8tion.jda.entities.TextChannel;
 
@@ -82,14 +84,14 @@ public class DiscordEndPoint extends EndPoint
     {
         if (!connected)
             throw new IllegalStateException("Cannot send message to disconnected EndPoint! EndPoint: " + this.toEndPointInfo().toString());
-        switch (message.getType())
+        if (message instanceof DiscordEndPointMessage)
         {
-            case DISCORD:
-                getChannel().sendMessage(message.getDiscordMessage());
-                break;
-            default:
-                for (String segment : this.divideMessageForSending(message.getMessage()))
-                    sendMessage(String.format("<%s> %s", message.getSenderName(), segment));
+            getChannel().sendMessage(((DiscordEndPointMessage) message).getDiscordMessage());
+        }
+        else if (message instanceof IrcEndPointMessage)
+        {
+            for (String segment : this.divideMessageForSending(message.getMessage()))
+                sendMessage(String.format("<%s> %s", message.getSenderName(), segment));
         }
     }
 }
